[aws-lib] move classes into module
