+++ aws-lib/src/main/scala/net/kemitix/s3thorp/aws/lib/CancellableMultiPartUpload.scala
@@ -1,0 +1,5 @@
+package net.kemitix.s3thorp.aws.lib
+
+final case class CancellableMultiPartUpload(
+  e: Throwable,
+  uploadId: String) extends Exception(e)
