+++ aws-lib/src/main/scala/net/kemitix/s3thorp/aws/lib/QuoteStripper.scala
@@ -1,0 +1,7 @@
+package net.kemitix.s3thorp.aws.lib
+
+trait QuoteStripper {
+
+  def stripQuotes: Char => Boolean = _ != '"'
+
+}
