+++ aws-lib/src/test/resources/net/kemitix/s3thorp/aws/lib/test-file-for-hash.txt
