+++ aws-lib/src/test/resources/net/kemitix/s3thorp/aws/lib/upload/subdir/leaf-file
