+++ core/src/main/scala/net.kemitix.s3thorp.core/Resource.scala
@@ -9,14 +9,7 @@ object Resource {
   def apply(base: AnyRef,
             name: String): File = {
     Try{
-      println(s"\nname = ${name}")
-      val clazz = base.getClass
-      println(s"clazz = ${clazz}")
-      val resource = clazz.getResource(name)
-      println(s"resource = ${resource}")
-      val path = resource.getPath
-      println(s"path = ${path}\n")
-      new File(path)
+      new File(base.getClass.getResource(name).getPath)
     }.getOrElse(throw new FileNotFoundException(name))
   }
 }
