--- src/main/scala/net/kemitix/s3thorp/awssdk/CancellableMultiPartUpload.scala
@@ -1,5 +1,0 @@
-package net.kemitix.s3thorp.awssdk
-
-final case class CancellableMultiPartUpload(
-  e: Throwable,
-  uploadId: String) extends Exception(e)
